+++ coinlib/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.0.0-rc.9
+
+- Update secp256k1 to 0.4.1
+- Update dart dependencies and FFI bindings
+
+## 2.0.0-rc.8
+
+Add windows build support
+
 ## 2.0.0-rc.7
 
 - Add `CoinUnit` class to convert between amount strings and satoshis.